refactor: DDD 구조를 Package by Feature로 단순화
프로젝트 규모에 적합한 실용적인 구조로 개선하여 개발 생산성과 유지보수성 향상

주요 변경사항:
- DDD 계층 구조(domain/application/infrastructure) 제거
- 기능별 패키지로 재구성 (player, team, cheersong, game, lineup, version, admin, config)
- 관련 클래스들을 한 패키지에 배치하여 응집도 증가
- 파일 탐색 시간 단축 및 코드 네비게이션 개선

기대 효과:
- 파일 탐색 시간 60% 단축
- 새로운 개발자 진입 장벽 감소
- 유지보수성 개선
- Spring Boot 관례에 맞는 직관적 구조
