+++ src/main/java/academy/cheerlot/cheersong/CheerSong.java
@@ -1,0 +1,27 @@
+package academy.cheerlot.cheersong;
+
+import jakarta.persistence.*;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "cheer_song")
+@Data
+@NoArgsConstructor
+public class CheerSong {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String title;
+    
+    @Column(columnDefinition = "TEXT")
+    private String lyrics;
+    
+    private String audioFileName;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "player_id")
+    private academy.cheerlot.player.Player player;
+}
