--- src/main/java/academy/cheerlot/domain/game/infrastructure/persistence/GameRepository.java
@@ -1,7 +1,0 @@
-package academy.cheerlot.domain.game.infrastructure.persistence;
-
-import academy.cheerlot.domain.game.domain.Game;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GameRepository extends JpaRepository<Game, String> {
-}
