+++ src/main/java/academy/cheerlot/game/GameRepository.java
@@ -1,0 +1,6 @@
+package academy.cheerlot.game;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface GameRepository extends JpaRepository<Game, String> {
+}
