+++ src/main/java/academy/cheerlot/lineup/LineupResponse.java
@@ -1,0 +1,12 @@
+package academy.cheerlot.lineup;
+
+import academy.cheerlot.player.PlayerDto;
+
+import java.util.List;
+
+public record LineupResponse(
+        String updated,
+        String opponent,
+        List<PlayerDto> players
+) {
+}
