+++ src/main/java/academy/cheerlot/player/PlayerDto.java
@@ -1,0 +1,21 @@
+package academy.cheerlot.player;
+
+public record PlayerDto(
+        Long id,
+        String name,
+        String backNumber,
+        String position,
+        String batsThrows,
+        String batsOrder
+) {
+    public static PlayerDto from(Player player) {
+        return new PlayerDto(
+                player.getId(),
+                player.getName(),
+                player.getBackNumber(),
+                player.getPosition(),
+                player.getBatsThrows(),
+                player.getBatsOrder()
+        );
+    }
+}
