+++ src/main/java/academy/cheerlot/team/TeamRepository.java
@@ -1,0 +1,6 @@
+package academy.cheerlot.team;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TeamRepository extends JpaRepository<Team, String> {
+}
